Tech web into typescript (#89116)
## About The Pull Request
Converts the tech web into typescript, removing uselocalstate, moving it
into separate files (huge!)

It should look and function exactly the same.
## Why It's Good For The Game
Just a massive file getting properly typed and sorted makes it easier
for this UI to digest & get the DMIcon treatment, which it needs, as it
might be our slowest UI

uselocalstate is also deprecated in favor of real hooks
## Changelog
N/A
